# Use scalar functions when less traffic