--- a/_gen/go.mod
+++ b/_gen/go.mod
@@ -1,0 +1,5 @@
+module github.com/minio/md5-simd/_gen
+
+go 1.14
+
+require github.com/mmcloughlin/avo v0.0.0-20210104032911-599bdd1269f4 // indirect