--- a/md5-util_amd64.go
+++ b/md5-util_amd64.go
@@ -12,6 +12,10 @@ type lane struct {
 	pos uint
 }
 
+type digest struct {
+	s [4]uint32
+}
+
 // Helper struct for generating number of rounds in combination with mask for valid lanes
 type maskRounds struct {
 	mask   uint64