… _gen/go.sum
… md5block_amd64.go